Move plugin under data silo resource
