+++ transcend/resource_data_silo_test.go
@@ -12,6 +12,24 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+func lookupDataSiloPlugin(t *testing.T, id string) []types.Plugin {
+	client := getTestClient()
+
+	var query struct {
+		Plugins struct {
+			Plugins []types.Plugin
+		} `graphql:"plugins(filterBy: { dataSiloId: $dataSiloId })"`
+	}
+	vars := map[string]interface{}{
+		"dataSiloId": graphql.String(id),
+	}
+
+	err := client.graphql.Query(context.Background(), &query, vars, graphql.OperationName("Plugins"))
+	assert.Nil(t, err)
+
+	return query.Plugins.Plugins
+}
+
 func lookupDataSilo(t *testing.T, id string) types.DataSilo {
 	client := getTestClient()
 
@@ -41,17 +59,18 @@ func prepareDataSiloOptions(t *testing.T, vars map[string]interface{}) *terrafor
 	return terraformOptions
 }
 
-func deployDataSilo(t *testing.T, terraformOptions *terraform.Options) types.DataSilo {
+func deployDataSilo(t *testing.T, terraformOptions *terraform.Options) (types.DataSilo, []types.Plugin) {
 	terraform.InitAndApplyAndIdempotent(t, terraformOptions)
 	assert.NotEmpty(t, terraform.Output(t, terraformOptions, "dataSiloId"))
 	silo := lookupDataSilo(t, terraform.Output(t, terraformOptions, "dataSiloId"))
-	return silo
+	plugin := lookupDataSiloPlugin(t, terraform.Output(t, terraformOptions, "dataSiloId"))
+	return silo, plugin
 }
 
 func TestCanCreateAndDestroyDataSilo(t *testing.T) {
 	options := prepareDataSiloOptions(t, map[string]interface{}{"title": t.Name()})
 	defer terraform.Destroy(t, options)
-	silo := deployDataSilo(t, options)
+	silo, _ := deployDataSilo(t, options)
 	assert.Equal(t, graphql.String(t.Name()), silo.Title)
 	assert.NotEmpty(t, terraform.Output(t, options, "awsExternalId"))
 }
@@ -59,7 +78,7 @@ func TestCanCreateAndDestroyDataSilo(t *testing.T) {
 func TestCanConnectAwsDataSilo(t *testing.T) {
 	options := prepareDataSiloOptions(t, map[string]interface{}{"skip_connecting": false})
 	defer terraform.Destroy(t, options)
-	silo := deployDataSilo(t, options)
+	silo, _ := deployDataSilo(t, options)
 	assert.Equal(t, graphql.String(t.Name()), silo.Title)
 	assert.NotEmpty(t, terraform.Output(t, options, "awsExternalId"))
 	assert.Equal(t, types.DataSiloConnectionState("CONNECTED"), silo.ConnectionState)
@@ -68,6 +87,7 @@ func TestCanConnectAwsDataSilo(t *testing.T) {
 func TestCanConnectDatadogDataSilo(t *testing.T) {
 	options := prepareDataSiloOptions(t, map[string]interface{}{
 		"skip_connecting": false,
+		"type":            "datadog",
 		"secret_context": []map[string]interface{}{
 			{
 				"name":  "apiKey",
@@ -84,15 +104,38 @@ func TestCanConnectDatadogDataSilo(t *testing.T) {
 		},
 	})
 	defer terraform.Destroy(t, options)
-	silo := deployDataSilo(t, options)
+	silo, _ := deployDataSilo(t, options)
 	assert.Equal(t, graphql.String(t.Name()), silo.Title)
 	assert.Equal(t, types.DataSiloConnectionState("CONNECTED"), silo.ConnectionState)
+}
+
+func TestCanConnectPlugin(t *testing.T) {
+	options := prepareDataSiloOptions(t, map[string]interface{}{
+		"skip_connecting": false,
+		"plugin_config": []map[string]interface{}{
+			{
+				"enabled":                    true,
+				"type":                       "DATA_POINT_DISCOVERY",
+				"schedule_frequency_minutes": 120,
+				// Schedule far in the future so that the test works for a long time
+				"schedule_start_at": "2122-09-06T17:51:13+0000",
+				"schedule_now":      false,
+			},
+		},
+	})
+	defer terraform.Destroy(t, options)
+	silo, plugins := deployDataSilo(t, options)
+	assert.Equal(t, graphql.String(t.Name()), silo.Title)
+	assert.Equal(t, types.DataSiloConnectionState("CONNECTED"), silo.ConnectionState)
+	assert.Len(t, plugins, 1)
+	assert.True(t, bool(plugins[0].Enabled))
+	assert.NotEmpty(t, plugins[0].ID)
 }
 
 func TestCanChangeTitle(t *testing.T) {
 	options := prepareDataSiloOptions(t, map[string]interface{}{"title": t.Name()})
 	defer terraform.Destroy(t, options)
-	silo := deployDataSilo(t, options)
+	silo, _ := deployDataSilo(t, options)
 	assert.Equal(t, graphql.String(t.Name()), silo.Title)
 
 	silo = deployDataSilo(t, prepareDataSiloOptions(t, map[string]interface{}{"title": t.Name() + "_2"}))
@@ -102,7 +145,7 @@ func TestCanChangeTitle(t *testing.T) {
 func TestCanChangeDescription(t *testing.T) {
 	options := prepareDataSiloOptions(t, map[string]interface{}{"description": t.Name()})
 	defer terraform.Destroy(t, options)
-	silo := deployDataSilo(t, options)
+	silo, _ := deployDataSilo(t, options)
 	assert.Equal(t, graphql.String(t.Name()), silo.Title)
 
 	silo = deployDataSilo(t, prepareDataSiloOptions(t, map[string]interface{}{"description": t.Name() + "_2"}))
@@ -112,7 +155,7 @@ func TestCanChangeDescription(t *testing.T) {
 func TestCanChangeUrl(t *testing.T) {
 	options := prepareDataSiloOptions(t, map[string]interface{}{"url": "https://some.webhook", "type": "server"})
 	defer terraform.Destroy(t, options)
-	silo := deployDataSilo(t, options)
+	silo, _ := deployDataSilo(t, options)
 	assert.Equal(t, graphql.String("https://some.webhook"), silo.URL)
 
 	silo = deployDataSilo(t, prepareDataSiloOptions(t, map[string]interface{}{"url": "https://some.other.webhook", "type": "server"}))
@@ -122,7 +165,7 @@ func TestCanChangeUrl(t *testing.T) {
 func TestCanChangeNotifyEmailAddress(t *testing.T) {
 	options := prepareDataSiloOptions(t, map[string]interface{}{"notify_email_address": "david@transcend.io"})
 	defer terraform.Destroy(t, options)
-	silo := deployDataSilo(t, options)
+	silo, _ := deployDataSilo(t, options)
 	assert.Equal(t, graphql.String("david@transcend.io"), silo.NotifyEmailAddress)
 
 	silo = deployDataSilo(t, prepareDataSiloOptions(t, map[string]interface{}{"notify_email_address": "mike@transcend.io"}))
@@ -132,7 +175,7 @@ func TestCanChangeNotifyEmailAddress(t *testing.T) {
 func TestCanChangeIsLive(t *testing.T) {
 	options := prepareDataSiloOptions(t, map[string]interface{}{"is_live": false})
 	defer terraform.Destroy(t, options)
-	silo := deployDataSilo(t, options)
+	silo, _ := deployDataSilo(t, options)
 	assert.Equal(t, graphql.Boolean(false), silo.IsLive)
 
 	silo = deployDataSilo(t, prepareDataSiloOptions(t, map[string]interface{}{"is_live": true}))
@@ -145,7 +188,7 @@ func TestCanChangeIsLive(t *testing.T) {
 func TestCanChangeOwners(t *testing.T) {
 	options := prepareDataSiloOptions(t, map[string]interface{}{"owner_emails": []string{"david@transcend.io"}})
 	defer terraform.Destroy(t, options)
-	silo := deployDataSilo(t, options)
+	silo, _ := deployDataSilo(t, options)
 	assert.Equal(t, graphql.String("david@transcend.io"), silo.Owners[0].Email)
 
 	silo = deployDataSilo(t, prepareDataSiloOptions(t, map[string]interface{}{"owner_emails": []string{"mike@transcend.io"}}))
@@ -161,7 +204,7 @@ func TestCanChangeHeaders(t *testing.T) {
 		},
 	}})
 	defer terraform.Destroy(t, options)
-	silo := deployDataSilo(t, options)
+	silo, _ := deployDataSilo(t, options)
 	assert.Equal(t, graphql.String("someHeader"), silo.Headers[0].Name)
 	assert.Equal(t, graphql.String("someHeaderValue"), silo.Headers[0].Value)
 
@@ -182,7 +225,7 @@ func TestCanCreatePromptAPersonSilo(t *testing.T) {
 		"outer_type": "coupa",
 	})
 	defer terraform.Destroy(t, options)
-	silo := deployDataSilo(t, options)
+	silo, _ := deployDataSilo(t, options)
 	assert.Equal(t, graphql.String("coupa"), silo.OuterType)
 	assert.Equal(t, graphql.String("promptAPerson"), silo.Type)
 	assert.Equal(t, graphql.Boolean(true), silo.Catalog.HasAvcFunctionality)
@@ -195,7 +238,7 @@ func TestCanSetPromptAPersonNotifyEmailAddress(t *testing.T) {
 		"notify_email_address": "not.real.email@transcend.io",
 	})
 	defer terraform.Destroy(t, options)
-	silo := deployDataSilo(t, options)
+	silo, _ := deployDataSilo(t, options)
 	assert.Equal(t, graphql.String("promptAPerson"), silo.Type)
 	assert.Equal(t, graphql.Boolean(true), silo.Catalog.HasAvcFunctionality)
 	assert.Equal(t, graphql.String("not.real.email@transcend.io"), silo.NotifyEmailAddress)
